refactor "SHOW KEY GENERATE ALGORITHM IMPLEMENTATIONS" module
