+++ features/sharding/distsql/parser/src/main/antlr4/imports/sharding/Keyword.g4
@@ -282,7 +282,3 @@ NONE
 IMPLEMENTATIONS
     : I M P L E M E N T A T I O N S
     ;
-
-GENERATE
-    : G E N E R A T E
-    ;
