+++ features/sharding/distsql/parser/src/main/antlr4/sharding/org/apache/shardingsphere/distsql/parser/autogen/ShardingDistSQLStatement.g4
@@ -47,6 +47,5 @@ execute
     | showUnusedShardingAuditors
     | countShardingRule
     | showShardingAlgorithmImplementations
-    | showShardingKeyGenerateAlgorithmImplementations
     ) SEMI_? EOF
     ;
