+++ parser/distsql/engine/src/main/antlr4/org/apache/shardingsphere/distsql/parser/autogen/KernelDistSQLStatement.g4
@@ -51,5 +51,6 @@ execute
     | lockCluster
     | unlockCluster
     | showServiceProviderImplementations
+    | showKeyGenerateAlgorithmImplementations
     ) SEMI_? EOF
     ;
